bot/requirements.txt
python-telegram-bot==20.3
aiofiles
openai
pydub
nest_asyncio
